Resign Windows binaries after promoting package
