+++ luna-manager/src/Luna/Manager/Command.hs
@@ -5,7 +5,6 @@ import Prologue
 import           Control.Monad.Raise
 import           Control.Monad.State.Layered
 import           Control.Monad.Trans.Resource       (MonadBaseControl)
-import           Luna.Manager.Command.CreatePackage (PackageConfig)
 import qualified Luna.Manager.Command.CreatePackage as CreatePackage
 import qualified Luna.Manager.Command.Develop       as Develop
 import           Luna.Manager.Command.Install       (InstallConfig)
@@ -16,6 +15,7 @@ import qualified Luna.Manager.Command.Promote       as Promote
 import qualified Luna.Manager.Command.Uninstall     as Uninstall
 import qualified Luna.Manager.Command.Version       as Version
 import           Luna.Manager.Component.Analytics   (MPUserData)
+import           Luna.Manager.Component.PackageConfig (PackageConfig)
 import           Luna.Manager.Component.Repository
 import           Luna.Manager.Shell.Shelly          (MonadSh, MonadShControl)
 import           Luna.Manager.System.Env
@@ -30,7 +30,7 @@ chooseCommand = do
         MakePackage opt -> evalDefHostConfigs @'[PackageConfig, EnvConfig, RepoConfig]                        $ CreatePackage.run opt
         Develop     opt -> evalDefHostConfigs @'[Develop.DevelopConfig, EnvConfig, PackageConfig, RepoConfig] $ Develop.run       opt
         NextVersion opt -> evalDefHostConfigs @'[EnvConfig, RepoConfig]                                       $ NextVersion.run   opt
-        Promote     opt -> evalDefHostConfigs @'[EnvConfig, RepoConfig]                                       $ Promote.run       opt
+        Promote     opt -> evalDefHostConfigs @'[EnvConfig, RepoConfig, PackageConfig]                        $ Promote.run       opt
         Uninstall       -> evalDefHostConfigs @'[InstallConfig, EnvConfig]                                    $ Uninstall.run
         Version         -> Version.run
         a               -> putStrLn $ "Unimplemented option: " ++ show a
